added pages to posts in profile
