+++ biostar/forum/templates/widgets/list_awards.html
@@ -1,9 +1,5 @@
 {% load forum_tags %}
 
-<div class="center-text">
-    {% pages objs=awards %}
-</div>
-
 {% if awards %}
 <div class="ui grid">
     {% for award in awards %}
