+++ biostar/forum/templatetags/forum_tags.py
@@ -18,11 +18,11 @@ from django.shortcuts import reverse
 from django.db.models import Q
 from django.utils.safestring import mark_safe
 from django.utils.timezone import utc
-from biostar.forum import markdown
+from django.core.paginator import Paginator
 
+from biostar.forum import markdown
 from biostar.accounts.models import Profile, Message
 from biostar.forum import const, auth
-from biostar.forum.auth import CachedPaginator
 from biostar.forum.models import Post, Vote, Award, Subscription
 
 User = get_user_model()
@@ -446,9 +446,12 @@ def search_bar(context, tags=False, users=False):
 
 @register.simple_tag
 def get_post_list(target, request, show=None):
-
+    """
+    Return post list belonging to a user
+    """
+    user = request.user
     page = request.GET.get("page", 1)
-    posts = Post.objects.filter(author=target)
+    posts = Post.objects.valid_posts(u=user, author=target)
 
     # Show a specific post listing.
     show_map = dict(questions=Post.QUESTION, tools=Post.TOOL, news=Post.NEWS,
@@ -458,12 +461,11 @@ def get_post_list(target, request, show=None):
     type_filter = show_map.get(show)
     posts = posts.filter(type=type_filter) if type_filter is not None else posts
 
-    posts = posts.select_related("root").select_related("author__profile", "lastedit_user__profile")
+    posts = posts.select_related("root").select_related( "author__profile", "lastedit_user__profile")
     posts = posts.order_by("-rank")
 
     # Cache the users posts add pagination to posts.
-    cache_key = f"{target.id}-KEY"
-    paginator = CachedPaginator(object_list=posts, cache_key=cache_key, per_page=settings.POSTS_PER_PAGE)
+    paginator = Paginator(object_list=posts, per_page=settings.POSTS_PER_PAGE)
     posts = paginator.get_page(page)
 
     return posts
@@ -528,7 +530,7 @@ def list_awards(context, target):
                                                               'badge').order_by("-date")
     page = request.GET.get('page', 1)
     # Create the paginator
-    paginator = CachedPaginator(object_list=awards, per_page=20)
+    paginator = Paginator(object_list=awards, per_page=20)
 
     # Apply the votes paging.
     awards = paginator.get_page(page)
